nexus: address child sizing mismatches
When a child fails to be added due to a different block size or invalid
size the returned error does not use the name fromm the bdev which is
now destroyed.

When adding a new child the size of the child has to be larger than the
minimum required.

The nexus bdev now takes into account the requested size when it was
created rather than just looking at its children sizes.
At the moment new children need to be as large as the smallest child,
this is due to the partition being done in relation to the smallest
child. We might want to consider doing it in relation to the nexus size
+ metadata.

Resolves CAS-247